Start liga merged con dev

+++ src/main/java/com/adondeband/back_end_adonde_band/jpa/jornada/JornadaJpaAdapter.java
@@ -2,6 +2,7 @@ package com.adondeband.back_end_adonde_band.jpa.jornada;
 
 import com.adondeband.back_end_adonde_band.dominio.jornada.Jornada;
 import com.adondeband.back_end_adonde_band.dominio.jornada.JornadaPort;
+import com.adondeband.back_end_adonde_band.dominio.jornada.JornadaId;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
@@ -26,7 +27,7 @@ public class JornadaJpaAdapter implements JornadaPort {
     }
 
     @Override
-    public List<Jornada> findById(int s) {
+    public List<Jornada> findById(Long s) {
         return jornadaJpaRepository.findById(s)
                 .stream()
                 .map(jornadaJpaMapper::toDomain)
